dcp-815 separate exceptions in metadata/exceptions.py
