+++ exporter/metadata/exceptions.py
@@ -1,0 +1,6 @@
+class MetadataParseException(Exception):
+    pass
+
+
+class MetadataException(Exception):
+    pass
