+++ exporter/terra/dcp_staging_client.py
@@ -14,8 +14,9 @@ from jsonschema.validators import validate
 from exporter.graph.experiment_graph import LinkSet
 from exporter.metadata.checksums import FileChecksums
 from exporter.metadata.datafile import DataFile
+from exporter.metadata.exceptions import MetadataParseException
 from exporter.schema import SchemaService
-from exporter.scratch import MetadataResource, MetadataParseException
+from exporter.scratch import MetadataResource
 from exporter.terra.gcs import GcsXferStorage, GcsStorage, Streamable, TransferJobSpec
 from exporter.utils import log_function_and_params
