+++ tests/exporter/test_metadata.py
@@ -6,8 +6,8 @@ from unittest.mock import patch, MagicMock, Mock
 from exporter import utils
 from exporter.metadata.checksums import FileChecksums
 from exporter.metadata.datafile import DataFile
-from exporter.scratch import MetadataResource, MetadataService, MetadataParseException, \
-    MetadataProvenance
+from exporter.metadata.exceptions import MetadataParseException
+from exporter.scratch import MetadataResource, MetadataService, MetadataProvenance
 
 
 class MetadataResourceTest(TestCase):
